set cmd_cnt member to max Functions
    cli_t * c1 = *cli;
    (*cli)->cmd_cnt = MAX_FUNCTIONS;
    c1->cmd_running = &c1->cmd_tbl[0];

+++ Core/Inc/olifex_cli_defs.h
@@ -41,7 +41,7 @@ typedef struct
 typedef struct
 {    
     println_func_ptr_t println; /* Function pointer to user defined println function.      */
-    size_t cmd_cnt;             /* Number of commands in cmd_tbl.  */
+    uint16_t cmd_cnt;             /* Number of commands in cmd_tbl.  */
     cmd_t  cmd_tbl[MAX_FUNCTIONS];
     cmd_t  * cmd_running;
 } cli_t;
